final module 1

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        float temperatureFahrenheit = temperatureCelsius*9/5+32;
+        float temperatureFahrenheit = ((float) temperatureCelsius)*9/5+32;
         System.out.println(temperatureFahrenheit);
     }
 }
